bot/requirements.txt
aiogram>=3.17.0
python-dotenv
httpx
